--- a/.idea/libraries/KotlinJavaScript.xml
+++ b/.idea/libraries/KotlinJavaScript.xml
@@ -3,6 +3,7 @@
     <CLASSES>
       <root url="jar://$KOTLIN_BUNDLED$/lib/kotlin-stdlib-js.jar!/" />
     </CLASSES>
+    <JAVADOC />
     <SOURCES>
       <root url="jar://$KOTLIN_BUNDLED$/lib/kotlin-stdlib-js-sources.jar!/" />
     </SOURCES>