--- a/src/asset/spr.rs
+++ b/src/asset/spr.rs
@@ -1,7 +1,7 @@
 use crate::asset::BinaryReader;
 
 pub struct SpriteFile {
-    pub frames: Vec<Frame>,
+    pub frames: Vec<SprFrame>,
 }
 
 pub enum SpriteType {
@@ -9,7 +9,7 @@ pub enum SpriteType {
     ABGR,
 }
 
-pub struct Frame {
+pub struct SprFrame {
     pub typ: SpriteType,
     pub width: usize,
     pub height: usize,
@@ -17,7 +17,7 @@ pub struct Frame {
 }
 
 impl SpriteFile {
-    pub(super) fn load(mut buf: BinaryReader, palette: Option<(usize, &Vec<u8>)>) -> Self {
+    pub(super) fn read_header(buf: &mut BinaryReader) -> (f32, usize, u16) {
         let header = buf.string(2);
         let version = buf.next_u8() as f32 / 10.0 + buf.next_u8() as f32;
         if header != "SP" {
@@ -26,6 +26,16 @@ impl SpriteFile {
 
         let indexed_frame_count = buf.next_u16() as usize;
         let rgba_frame_count = if version > 1.1 { buf.next_u16() } else { 0 };
+        return (version, indexed_frame_count, rgba_frame_count);
+    }
+
+    pub(super) fn load(
+        mut buf: BinaryReader,
+        palette: Option<(usize, Vec<u8>)>,
+        version: f32,
+        indexed_frame_count: usize,
+        rgba_frame_count: u16,
+    ) -> Self {
         let indexed_frames = if version < 2.1 {
             SpriteFile::read_indexed_frames(&mut buf, indexed_frame_count)
         } else {
@@ -34,13 +44,15 @@ impl SpriteFile {
 
         let rgba_frames = SpriteFile::read_rgba_frames(&mut buf, rgba_frame_count);
 
-        let default_palette = if version > 1.0 {
-            buf.skip(((buf.len() - 1024) - buf.tell()) as u32);
-            buf.next(1024)
-        } else {
-            Vec::new()
+        let palette = {
+            let default_palette = if version > 1.0 {
+                buf.skip(((buf.len() - 1024) - buf.tell()) as u32);
+                buf.next(1024)
+            } else {
+                Vec::new()
+            };
+            palette.map(|it| it.1).unwrap_or(default_palette)
         };
-        let palette = palette.map(|it| it.1).unwrap_or(&default_palette);
 
         let mut frames = Vec::with_capacity(indexed_frames.len() + rgba_frames.len());
 
@@ -54,12 +66,12 @@ impl SpriteFile {
         SpriteFile { frames }
     }
 
-    fn read_indexed_frames(buf: &mut BinaryReader, indexed_frame_count: usize) -> Vec<Frame> {
+    fn read_indexed_frames(buf: &mut BinaryReader, indexed_frame_count: usize) -> Vec<SprFrame> {
         (0..indexed_frame_count)
             .map(|_i| {
                 let width = buf.next_u16();
                 let height = buf.next_u16();
-                Frame {
+                SprFrame {
                     typ: SpriteType::PAL,
                     width: width as usize,
                     height: height as usize,
@@ -69,7 +81,7 @@ impl SpriteFile {
             .collect()
     }
 
-    fn to_rgba(frame: Frame, pal: &Vec<u8>) -> Frame {
+    fn to_rgba(frame: SprFrame, pal: &Vec<u8>) -> SprFrame {
         let mut buf = Vec::<u8>::with_capacity((frame.width * frame.height * 4) as usize);
         for y in 0..frame.height {
             for x in 0..frame.width {
@@ -80,14 +92,17 @@ impl SpriteFile {
                 buf.push(if idx1 != 0 { 255 } else { 0 }); // a
             }
         }
-        Frame {
+        SprFrame {
             typ: SpriteType::ABGR,
             data: buf,
             ..frame
         }
     }
 
-    fn read_indexed_frames_rle(buf: &mut BinaryReader, indexed_frame_count: usize) -> Vec<Frame> {
+    fn read_indexed_frames_rle(
+        buf: &mut BinaryReader,
+        indexed_frame_count: usize,
+    ) -> Vec<SprFrame> {
         (0..indexed_frame_count)
             .map(|_i| {
                 let width = buf.next_u16();
@@ -108,7 +123,7 @@ impl SpriteFile {
                         }
                     }
                 }
-                Frame {
+                SprFrame {
                     typ: SpriteType::PAL,
                     width: width as usize,
                     height: height as usize,
@@ -118,7 +133,7 @@ impl SpriteFile {
             .collect()
     }
 
-    fn read_rgba_frames(buf: &mut BinaryReader, rgba_frame_count: u16) -> Vec<Frame> {
+    fn read_rgba_frames(buf: &mut BinaryReader, rgba_frame_count: u16) -> Vec<SprFrame> {
         (0..rgba_frame_count)
             .map(|_i| {
                 let width = buf.next_u16();
@@ -126,7 +141,7 @@ impl SpriteFile {
                 let mut data = buf.next(width as u32 * height as u32 * 4);
                 // it seems ABGR sprites are stored upside down
                 data.reverse();
-                Frame {
+                SprFrame {
                     typ: SpriteType::ABGR,
                     width: width as usize,
                     height: height as usize,