--- a/src/systems/mod.rs
+++ b/src/systems/mod.rs
@@ -67,6 +67,7 @@ pub struct Sprites {
     pub falcon: SpriteResource,
     pub stun: SpriteResource,
     pub timefont: SpriteResource,
+    // TODO: make it array
     pub character_sprites: HashMap<JobSpriteId, [[SpriteResource; 2]; 2]>,
     pub mounted_character_sprites: HashMap<JobId, [SpriteResource; 2]>,
     pub head_sprites: [Vec<SpriteResource>; 2],