Sprites and models are loaded on a background thread asynchronously.
    - old, debug mode: 24s
    - old, release mode: ~3.5s (max value 6.1s)
    - new, debug mode: 7.6s
    - new, release mode: 2.7s (fairly stable, no fluctuations)